pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.40",
    "pydantic>=2.5",
    "polars>=0.20",
    "duckdb>=0.10",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "statsmodels>=0.14",
    "scikit-learn>=1.4",
    "fredapi>=0.5",
    "yfinance>=0.2",
    "sentence-transformers>=2.5",
    "faiss-cpu>=1.7",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "huggingface-hub>=0.22",
    "datasets>=2.18",
    "matplotlib>=3.10.9",
]

[tool.ruff]
line-length = 100

[tool.uv]
package = false

